+++ imod/tests/test_mf6/test_mf6_simulation.py
@@ -1,4 +1,5 @@
 import os
+import sys
 from unittest import mock
 from unittest.mock import MagicMock
 
@@ -42,6 +43,66 @@ def test_circle_roundtrip(circle_model, tmpdir_factory):
     roundtrip(circle_model, tmpdir_factory, "circle")
 
 
+@pytest.mark.usefixtures("circle_model")
+@pytest.mark.skipif(sys.version_info < (3, 7), reason="capture_output added in 3.7")
+def test_simulation_open_head(circle_model, tmp_path):
+    simulation = circle_model
+
+    # Should throw error when model not run yet.
+    with pytest.raises(RuntimeError):
+        simulation.open_head()
+
+    modeldir = tmp_path / "circle"
+    simulation.write(modeldir)
+    simulation.run()
+
+    head = simulation.open_head()
+
+    assert isinstance(head, xu.UgridDataArray)
+    assert head.dims == ("time", "layer", "mesh2d_nFaces")
+    assert head.shape == (1, 2, 216)
+
+
+@pytest.mark.usefixtures("circle_model")
+@pytest.mark.skipif(sys.version_info < (3, 7), reason="capture_output added in 3.7")
+def test_simulation_open_flow_budget(circle_model, tmp_path):
+    simulation = circle_model
+
+    # Should throw error when model not run yet.
+    with pytest.raises(RuntimeError):
+        simulation.open_flow_budget()
+
+    modeldir = tmp_path / "circle"
+    simulation.write(modeldir, binary=False, use_absolute_paths=True)
+    simulation.run()
+
+    budget = simulation.open_flow_budget()
+
+    assert isinstance(budget, dict)
+    assert sorted(budget.keys()) == [
+        "chd",
+        "flow-horizontal-face",
+        "flow-horizontal-face-x",
+        "flow-horizontal-face-y",
+        "flow-lower-face",
+    ]
+    assert isinstance(budget["chd"], xu.UgridDataArray)
+
+
+@pytest.mark.usefixtures("circle_model")
+@pytest.mark.skipif(sys.version_info < (3, 7), reason="capture_output added in 3.7")
+def test_simulation_open_concentration_fail(circle_model, tmp_path):
+    """No transport model is assigned, so should throw error when opening concentrations"""
+    simulation = circle_model
+
+    modeldir = tmp_path / "circle"
+    simulation.write(modeldir, binary=False, use_absolute_paths=True)
+    simulation.run()
+
+    with pytest.raises(ValueError):
+        simulation.open_concentration()
+
+
 @pytest.fixture(scope="function")
 def setup_simulation():
     simulation = imod.mf6.Modflow6Simulation("TestSimulation")
@@ -57,6 +118,21 @@ def setup_simulation():
 
 
 class TestModflow6Simulation:
+    def test_write_sets_directory(self, tmp_path, setup_simulation):
+        # Arrange.
+        simulation = setup_simulation
+
+        # Assert
+        # Should be None upon initialization
+        assert simulation.directory is None
+
+        # Act.
+        simulation.write(tmp_path)
+
+        # Assert.
+        assert simulation.directory is not None
+        assert simulation.directory == tmp_path
+
     def test_write_with_default_arguments_writes_expected_files(
         self, tmp_path, setup_simulation
     ):